+++ src/components/BotShareCompleted.js
@@ -2,6 +2,8 @@ import React from 'react';
 import {View, Text} from 'react-native';
 import {k, width, height} from './Global';
 import {Actions} from 'react-native-router-native';
+import Avatar from "./Avatar";
+import profileFactory from '../factory/profileFactory';
 
 export default class extends React.Component {
     componentDidMount() {
@@ -9,15 +11,33 @@ export default class extends React.Component {
     }
 
     render() {
+        const profile = profileFactory.create(this.props.user);
         return <View style={{
-            width, height, justifyContent: 'center',
-            alignItems: 'center', backgroundColor: 'rgba(254,92,108,0.9)'
+            width,
+            height,
+            backgroundColor: 'transparent',
+            paddingTop: 175 * k,
+            paddingRight: 82 * k,
+            paddingLeft: 82 * k,
         }}>
-            <View style={{paddingTop: 294 * k, paddingRight: 80 * k, paddingLeft: 80 * k, paddingBottom: 274 * k}}>
-                <Text style={{fontFamily: 'Roboto-Medium', textAlign: 'center', color: 'white', fontSize: 22 * k}}>Awesome!{'\n\n'}
-                    <Text style={{fontFamily: 'Roboto-Regular', fontSize: 20 * k}}>You just shared this bot{'\n'}
-                        with @{this.props.profile.handle}</Text>
-                </Text>
+            <View style={{
+                backgroundColor: 'rgba(255,255,255,0.9)',
+                height: 80 * k,
+                borderRadius: 74 * k,
+                flexDirection: 'row'
+            }}>
+                {this.props.number > 1 &&
+                <View style={{position: 'absolute', top: 20 * k, left: 50 * k}}><Avatar size={40 * k}
+                                                                                        text={"+" + (this.props.number - 1)}/></View>}
+                <View style={{position: 'absolute', top: 20 * k, left: 20 * k}}><Avatar size={40 * k}
+                                                                                        profile={profile}/></View>
+                <View style={{position: 'absolute', top: 20 * k, left: 110 * k, width: 95 * k}}>
+                    <Text style={{color: 'rgb(63,50,77)', fontSize: 14 * k, fontFamily: 'Roboto-Medium'}}>
+                        Bot Shared <Text style={{fontFamily: 'Roboto-Regular'}}>with </Text>{this.props.number}
+                        friend{this.props.number > 1 ? 's' : ''}
+                    </Text>
+                </View>
+
             </View>
         </View>
     }